+++ flutter/lib/src/sentry_flutter_options.dart
@@ -4,6 +4,7 @@ import 'package:flutter/widgets.dart';
 
 import 'binding_wrapper.dart';
 import 'renderer/renderer.dart';
+import 'screenshot/sentry_screenshot_quality.dart';
 
 /// This class adds options which are only availble in a Flutter environment.
 /// Note that some of these options require native Sentry integration, which is
@@ -144,7 +145,7 @@ class SentryFlutterOptions extends SentryOptions {
 
   /// Enable scope sync from Java to NDK.
   /// Only available on Android.
-  bool enableNdkScopeSync = false;
+  bool enableNdkScopeSync = true;
 
   /// Enable auto performance tracking by default.
   bool enableAutoPerformanceTracing = true;
@@ -162,6 +163,9 @@ class SentryFlutterOptions extends SentryOptions {
   /// runApp(SentryScreenshotWidget(child: App()));
   /// The [SentryScreenshotWidget] has to be the root widget of the app.
   bool attachScreenshot = false;
+
+  /// The quality of the attached screenshot
+  SentryScreenshotQuality screenshotQuality = SentryScreenshotQuality.high;
 
   /// Enable or disable automatic breadcrumbs for User interactions Using [Listener]
   ///